Implemented formatters via the FormattedBy annotation.
Users can now specify the formatter for their data by using the
FormattedBy annotation.  This will tell Snitch what code to use
when formatting a specific object for user output.  For example:

@FormattedBy(MyFormatter.class)
class MyData {
  ...
}

class MyFormatter implements Formatter<MyData> {
  public void format(MyData obj, JsonWriter writer) throws IOException {
    ...
  }
}

This will cause Snitch to always use MyFormatter whenever it tries
to output instances of MyData.

This approach is not optimal, if MyData and MyFormatter are owned by
different groups then it's not always possible to get the proper
annotation onto the MyData class.  In the future Snitch should provide
an API for the implementers of MyFormatter to specify the binding
so that the authors of MyData are not involved.