Only process classes in the context of the class loader that owns them.
It (seems) possible for multiple class loaders to get a handle to a
class, even though only 1 of them is the "official" class loader for
it (e.g. cls.getClassLoader() == loader).

In this case we should only process classes when we process the loader
that owns them.  This will ensure that no unintentional leaking of
classes across class loaders happen.